Removed interface AgencyUserManager
